eds: log a warning when config contains duplicate localities
Currently if EDS returns multiple localities with the same Locality
proto they will be treated as the same locality. This is fine in most
cases but will cause surprising behavior under certain circumstances
(e.g. when the weights are different). This PR adds a log line to help
debugging these cases.

Signed-off-by: Snow Pettersen <snowp@squareup.com>