upstream: optimize membership update cb for eds
This makes it possible for a Cluster to opt out of the behavior that any
changes to the HostSet triggers a membership update callback, and
updates EdsClusterImpl to make use of this behavior.

Since all priorities are updated at the same time with EDS, we can keep track of
what hosts were actually added/removed vs just moved between priorities.
This avoids connection draining and reseting the active health checker
of these moved endpoints.

This also works towards #4280 by pulling membership updates out of the
HotSet, which will similarily rely on triggering membership callbacks
outside of the HostSet.

Signed-off-by: Snow Pettersen <snowp@squareup.com>