upstream: ignore endpoints with unknown health check status
Prior to this PR, when active health checking was enabled and hosts
are added through EDS/DNSthey are initialized as unhealthy pending
the first active health check result. This would cause the host
health of each HostSet to be scewed, potentially causing priority
spillover and panic mode.

This PR makes it so that these endpoints are not sent to the worker
threads, effectively excluding them from consideration until they've
been health checked at least once.

Signed-off-by: Snow Pettersen <snowp@squareup.com>